requirements.txt
pytz==2017.2
django==1.11.*
psycopg2-binary==2.8.*
python-dotenv==0.19.2
environs==9.5.0
dj-database-url==0.5.0